Decompose full node simulator HTTP module
