+++ crates/chia-sdk-test/src/full_node_simulator_http/state.rs
@@ -1,0 +1,5 @@
+use std::sync::{Arc, Mutex};
+
+use crate::FullNodeSimulator;
+
+pub type SharedSimulator = Arc<Mutex<FullNodeSimulator>>;
